Move Announcement to private folder
